Revert "Android 8.0, 8.1 and 9.0 Support (#2553)"
This reverts commit 17cdeae858b4c8a88b408fe743e6d8f2bdcbb72b.
