+++ examples/app.js
@@ -29,7 +29,6 @@ import RTL from './rtl'
 import ReadOnly from './read-only'
 import RichText from './rich-text'
 import SearchHighlighting from './search-highlighting'
-import Composition from './composition'
 import InputTester from './input-tester'
 import SyncingOperations from './syncing-operations'
 import Tables from './tables'
@@ -44,7 +43,6 @@ import Mentions from './mentions'
 const EXAMPLES = [
   ['Check Lists', CheckLists, '/check-lists'],
   ['Code Highlighting', CodeHighlighting, '/code-highlighting'],
-  ['Composition', Composition, '/composition/:subpage?'],
   ['Embeds', Embeds, '/embeds'],
   ['Emojis', Emojis, '/emojis'],
   ['Forced Layout', ForcedLayout, '/forced-layout'],
@@ -264,13 +262,11 @@ export default class App extends React.Component {
       <Switch>
         {EXAMPLES.map(([name, Component, path]) => (
           <Route key={path} path={path}>
-            {({ match }) => (
-              <div>
-                <ExampleContent>
-                  <Component params={match.params} />
-                </ExampleContent>
-              </div>
-            )}
+            <div>
+              <ExampleContent>
+                <Component />
+              </ExampleContent>
+            </div>
           </Route>
         ))}
         <Redirect from="/" to="/rich-text" />
@@ -294,7 +290,7 @@ export default class App extends React.Component {
         </TabButton>
         <Switch>
           {EXAMPLES.map(([name, Component, path]) => (
-            <Route key={path} path={path}>
+            <Route key={path} exact path={path}>
               <ExampleTitle>{name}</ExampleTitle>
             </Route>
           ))}
