+++ package.json
@@ -93,10 +93,8 @@
     "bootstrap": "lerna bootstrap && yarn build",
     "build": "rollup --config ./support/rollup/config.js",
     "build:production": "cross-env NODE_ENV=production rollup --config ./support/rollup/config.js && cross-env NODE_ENV=production webpack --config support/webpack/config.js",
-    "build:clean-fork": "rm ./build/CNAME",
     "clean": "lerna run clean && rm -rf ./node_modules ./dist ./build",
     "gh-pages": "gh-pages --dist ./build",
-    "gh-pages:fork": "npm-run-all build:production build:clean-fork gh-pages",
     "lint": "yarn lint:eslint && yarn lint:prettier",
     "lint:eslint": "eslint benchmark packages/*/src packages/*/test examples/*/*.js examples/dev/*/*.js",
     "lint:prettier": "prettier --list-different '**/*.{md,json,css}'",
