+++ packages/slate-react/src/plugins/dom.js
@@ -1,6 +1,3 @@
-import { IS_ANDROID } from 'slate-dev-environment'
-import AndroidPlugin from './android'
-import DebugPlugin from './debug'
 import AfterPlugin from './after'
 import BeforePlugin from './before'
 
@@ -13,15 +10,9 @@ import BeforePlugin from './before'
 
 function DOMPlugin(options = {}) {
   const { plugins = [] } = options
-  // Add Android specific handling separately before it gets to the other
-  // plugins because it is specific (other browser don't need it) and finicky
-  // (it has to come before other plugins to work).
-  const beforeBeforePlugins = IS_ANDROID
-    ? [AndroidPlugin(), DebugPlugin('slate:debug')]
-    : []
   const beforePlugin = BeforePlugin()
   const afterPlugin = AfterPlugin()
-  return [...beforeBeforePlugins, beforePlugin, ...plugins, afterPlugin]
+  return [beforePlugin, ...plugins, afterPlugin]
 }
 
 /**
